requirements.txt
numpy>=1.24
pandas>=2.0
SimpleITK>=2.3
PyYAML>=6.0
matplotlib>=3.7
pytest>=7.0
